render.yaml update
Have read up, and realise the utility of render.yaml is that you can track changes to your infrastructure over time as they match your commits. You can include existing services in the blueprint, no need to delete in the dashboard and recreate. Am attempting to add the existing database into this config file to see whether render will pick up on this.