render.yaml updated
The hash value used internally by render may not be what needs to be supplied for the databaseName option - I got an error complaining about this when running the blueprint. This commit opts for dynamite_db here instead, which is a value that I specified as a name for the database when I was creating it.