[FEAT] students.age column added (simple for now)
The existing booking form for dynamite includes an age field. Age is
easier to fill in than birthday. Birthday is more authoritative, but
may come across as intrusive. For now, age is a simple attribute that
won't dynamically change year on year. It is also not tied to a
birthday. At a later date, we can tie the two together if it would add
utility to the app; i.e. specify the age range (e.g. 7-8 years old)
update on a year anniversary following form submission (e.g. 8-9).
Upon provision of birthday, set age date accordingly. For now this
isn't necessary.