[REFAC] lesson#formatted_start_time for #to_s
Needed access to Lesson#to_s that references the start time of the
lesson in a formatted manner - the driver to do this is
ensure_not_referenced which uses the to_s of the model instance it is
acting on. This will be relevant for lessons instances once we build
out the lesson_members join table.