[CHG] Tidy Student validations
No longer require gender, but, when given, that it must be one of the
specified input values. Specify some new error messages for a couple
other validations.

Not requiring gender because the form submission from bookings won't
query child gender ~ we want to make it lowkey and simple to book.
Therefore, not going to require student instances require gender input
to pass validation.