[CHG] CSS button selector to target delete only
Had a style to put a margin-top on buttons so that delete buttons
would appear further down the page, out of the way, less prone to
being accidentally pressed. That style was applied to all buttons
though, this change improves the selector targeting to only apply the
style to delete buttons.