[REFAC] Tutor#active_tutors scope method
Experimenting with using scope methods. This seems to be a clean and
suitable place to define the query that retrieve active tutors.
I don't know where else would be suitable to use scope methods, and
where I should just rely on queries inplace in my controllers, or
perhaps in helper methods.