[CHG] Remove unnecessary show-student div class
This class, attached to a div that wrapped all of the content except
the heading in the students/show.html.erb view, was not contributing
anything. It was born from an idea that there might be specific styles
we'd want to apply to the student version of the show-instance div.
(show-instance is a class applied to the same div. At the moment it is
just providing padding.) A class that served as a hook to hang student
specific show styles might come in handy, but only if I have those
styles to hang from it, which I don't right now.