[BUG] Render _new_lesson_member as temp, nt partial
In the #create action, the failure clause of the conditional in the
respond_to block was rendering the partial *as* a partial. Should be
as a template for now. The reason for using a partial is that I plan
to insert it into the lessons/show template at some stage. Not yet
though.