[CHG] #decomission_old_version skips validation
Previously, a newly saved version of a contact would prevent the old
version from having it's valid_until attribute updated due to a
validation clash with the new version; i.e. the old tutor's phone and
email would not register as unique since the new version saved in the
database shared the same email / phone.

Skipping validation with the #update_attribute method.