[CHG] Remove entity_id's from application
The idea behind entity_id was that a set of records in a single table
might pertain to the same entity, with each distinct record in the set
storing a separate version of the entity.
This didn't end up playing out well when it came time to form
foreign key associations between records in other tables with the
records in a table that represent an entity ~ databases are designed
in such a way that foreign keys will always share a singular unique
key on another table.
The approach I was aiming for would have required that all of a set of
records in the referenced table be potentially referable by the same
entity_id. I hadn't really thought about what would designate the
specific referenced record to associate when the time came to use it;
one approach might have been to filter to the most up-to-date version
of the referenced records for use when the association was being
leveraged by the application.

At this stage I've stripped entity_ids entirely out of the application
- helpers, hook methods, the entity_id_references table, and the
  columns on tutors and contacts that stored the entity_id.

If I still want to be able to store versions of entities in the future
I might aim for an archiving approach, where an update would trigger
that the current versions data was stored in an archive table, with a
foreign key reference to the version that was updated. For now though,
I'll leave this as a question to pursue with Simon regarding how much
history we want to retain.

I'll still hold onto decomissioning over deleting for now.
The next commit will alter the update behaviours for all relevant
controllers to ensure that the original entity is updated, rather than
beind decomissioned with a new entity saved in it's place.