[CHG] Drop lessons.avl; use LM counter_cch instead
counter_cache may or may not have functional benefits. It certainly
seems to be working with the grain, which is a good thing to do with
Rails.

LessonMembersController #increment/#decrement_lesson_members_count
-> Keeps the lesson_members_count up to date for Lesson instances.

For now, we're building the functionality of the app without any turbo
streams. This is so that there is a fallback option for browsers that
won't activate turbo_stream. I can come back later to tidy this up.