[BUILD] Bookings spike - #create action and helpers
Quite a few helpers here and there. In this spike I was mapping the
form inputs to the attributes of a Student and Contact model manually,
using class methods defined on Booking. Also ascertaining the
lesson_id from the params, and the region_id from a join query on the
venue of the lesson with that id.

I'm going to go for the Rails convention after this. Will involve a
refactor, using accepts_nested_attributes_for declarations in the
Booking model, as well as some fields_for helpers in bookings/_form
to specify which inputs are for which derivate model instances
(contact or student)