[BUILD] Booking Spike
At this stage the form is failing when the student or contact is
invalid, and displaying the error messages derived from those
associations. This is behaviour that we want, but there are still
several issues with the current state of affairs:

1. Values entered in fields_for inputs are not persisted when the
   submission fails. This is because they are not values stored on the
   @booking model itself.

2. The error messages displayed are derived from the associated
   models; it would be better if we could customize them specifically
   for the form.
   - A subissue with the error messages being derived is that I've had
     to set the inputs as first_name and last_name; the error message
     that propogates from a not entered full_name is tied to the
     first_name not being supplied. Ideally though, for a better UX,
     the user can supply the name as a full_name, however they want,
     and if they don't supply an input, then the error message will
     pertain to this.

I am next going to explore the gem solutions on offer that might allow
for inputs tied to the booking form that are used to construct the
associated models, but that themselves are linked to the booking. This
sounds similar to my original implementation, and I wouldn't be
surprised if it were, and if it circumvented the whole
'accepts_nested_attributes_for' shenanigans.

Some options worth looking into are simple_form and cocoon. There are
good videos exploring these gems available on the 'deanin' youtube
channel.